[Refactor] Refactor the Database source file directory
