+++ src/renderer/views/components/Database/Delete/index.vue
