+++ src/renderer/views/components/table/Column/Info/index.vue
